requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1757067186903/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
distro==1.9.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
faiss-cpu==1.7.4
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.50.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==7.2.1
ipykernel @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_ipykernel_1761567969/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1759151815/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kagglehub==0.3.13
kiwisolver==1.4.9
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.23
langchain-openai==0.0.2
langsmith==0.0.87
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.5
nltk==3.9.2
numpy==1.26.3
openai==1.10.0
packaging==23.2
pandas==2.1.4
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
plotly==6.3.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
protobuf==4.25.8
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896323688/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1758103941054/work
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==13.9.4
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==2.3.1
sentencepiece==0.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
streamlit==1.30.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
websockets==15.0.1
yarl==1.22.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
